+++ docs/changes.rst
@@ -28,6 +28,7 @@ Release history
 - feat: new :meth:`wxc_sdk.licenses.LicensesApi.assigned_users`, :meth:`wxc_sdk.licenses.LicensesApi.assign_licenses_to_users`
 - feat: new :attr:`wxc_sdk.workspaces.Workspace.location_id`
 - feat: call record events :class:`wxc_sdk.events.EventResource`, :class:`wxc_sdk.events.EventData`
+- feat: new API: :class:`wxc_sdk.telephony.dect_devices.DECTDevicesApi`
 
 1.16.1
 ------